Agregue lógica al cancelar el cambio de estado de pedido

+++ el-buen-sabor/src/context/ImagenContext.tsx
@@ -2,7 +2,6 @@ import { ReactNode, createContext, useState } from "react";
 import { IImagen, ImageFile } from "../types/Articulos/ImagenArticulo";
 import Swal from "sweetalert2";
 import { ImagenesService } from "../services/ImagenesService";
-import { extractPublicId } from "cloudinary-build-url";
 import { imageFileVacio } from "../types/TiposVacios";
 
 interface ImageContextType {
@@ -125,26 +124,31 @@ export function ImageContextProvider({ children } : { children: ReactNode }) {
         return "";
      };
 
-    //delete toDeleteImages from cloudinary
     const deleteImages = () => {
         let service = new ImagenesService(objUrl);
-        
-        console.log(toDeleteImage.imagen.url);
-        let regex = /\/([0-9a-f]{32})/;
-        let match2 = toDeleteImage.imagen.url.match(regex);
-        //extraigo el ID publico de la URL de la imagen
-        const match = extractPublicId(toDeleteImage.imagen.url);
-        console.log(match[1])
-        if(match2) console.log(match2[1])
-        if(match){
-            const publicId = match[1];
-            try {
-                //Realizar la peticion post para eliminar la imagen
-                service.delete({id: toDeleteImage.imagen.id, publicId: publicId});
-            } catch (error) {
-                throw new Error();
+            console.log(toDeleteImage.imagen.url);
+    
+            // Extraigo el ID público de la URL de la imagen
+            const match = extraerPublicId(toDeleteImage.imagen.url);
+            console.log(match);
+    
+            if (match && match[1]) {
+                const publicId = match[1];
+                console.log(match[1])
+                try {
+                    // Realizar la petición POST para eliminar la imagen
+                     service.delete({ id: toDeleteImage.imagen.id, publicId: publicId });
+                } catch (error) {
+                    console.error("Error al eliminar la imagen:", error);
+                }
+            } else {
+                console.error("No se pudo extraer el publicId de la URL:", toDeleteImage.imagen.url);
             }
-        }
+    };
+
+    const extraerPublicId = (url: string): RegExpMatchArray | null => {
+        const regex = /\/upload\/(.+?)$/;
+        return url.match(regex);
     };
 
     //empty the states
